+++ packages/app/src/actions/index.js
@@ -21,7 +21,11 @@ import { VARIANT_SUCCESS } from '../components/Snackbar/Snackbar'
 
 import { sGetCurrent } from '../reducers/current'
 import { sGetVisualization } from '../reducers/visualization'
-import { sGetRootOrgUnit, sGetRelativePeriod } from '../reducers/settings'
+import {
+    sGetRootOrgUnit,
+    sGetRelativePeriod,
+    sGetSettingsDigitGroupSeparator,
+} from '../reducers/settings'
 
 import history from '../modules/history'
 import { getVisualizationFromCurrent } from '../modules/visualization'
@@ -31,6 +35,7 @@ import {
     GenericServerError,
     VisualizationNotFoundError,
 } from '../modules/error'
+import { getDisplayNameByVisType } from '@dhis2/analytics'
 
 export {
     fromVisualization,
@@ -113,8 +118,15 @@ export const clearVisualization = (dispatch, getState, error = null) => {
 
     const rootOrganisationUnit = sGetRootOrgUnit(getState())
     const relativePeriod = sGetRelativePeriod(getState())
+    const digitGroupSeparator = sGetSettingsDigitGroupSeparator(getState())
 
-    dispatch(fromUi.acClear({ rootOrganisationUnit, relativePeriod }))
+    dispatch(
+        fromUi.acClear({
+            rootOrganisationUnit,
+            relativePeriod,
+            digitGroupSeparator,
+        })
+    )
 }
 
 export const tDoRenameVisualization = ({ name, description }) => (
@@ -188,9 +200,16 @@ export const tDoSaveVisualization = ({ name, description }, copy) => async (
             delete visualization.id
         }
 
-        if (name) {
-            visualization.name = name
-        }
+        visualization.name =
+            name ||
+            i18n.t('Untitled {{visualizationType}} visualization, {{date}}', {
+                visualizationType: getDisplayNameByVisType(visualization.type),
+                date: new Date().toLocaleDateString(undefined, {
+                    year: 'numeric',
+                    month: 'short',
+                    day: '2-digit',
+                }),
+            })
 
         if (description) {
             visualization.description = description